+++ ui/src/Pages/BirdiesPage.js
@@ -13,7 +13,7 @@ import {
     fetchBirdieBatchById,
     fetchInventoryAdjustmentsForBatch,
     fetchSessionUsageForBirdieBatch
-} from '../firebaseService';
+} from '../services/firebaseService';
 import { collection, getDocs, query, orderBy } from 'firebase/firestore';
 import { Link } from 'react-router';
